+++ database/migrations/2025_05_10_170803_create_dishes_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('category')->default('main_dish');
             $table->string('imagePath')->nullable();
             $table->boolean('track_stock')->default(true);
-            $table->float('stock')->default(0); // Only used if track_stock = true
+            $table->integer('stock')->default(0); // Only used if track_stock = true
             $table->boolean('available')->default(true);
             $table->boolean('isActive')->default(true);
             $table->timestamps();
